视频新闻

+++ src/main/java/com/yuqiaotech/zsnews/controller/UploadController.java
@@ -41,12 +41,13 @@ public class UploadController extends BaseController
     
     @RequestMapping(value = "/uploadFile", method = RequestMethod.POST)
     @ResponseBody
-    public String uploadFile(@RequestParam(value="file",required=false) MultipartFile file, @RequestParam("objectId") String objectId,
-        @RequestParam("objectType") String objectType,@RequestParam(value="imgBase64",required=false) String imgBase64
-            ,@RequestParam(value="imgBase64Name",required=false) String imgBase64Name, HttpServletRequest request)
+    public String uploadFile(@RequestParam(value = "file", required = false) MultipartFile file,
+        @RequestParam("objectId") String objectId, @RequestParam("objectType") String objectType,
+        @RequestParam(value = "imgBase64", required = false) String imgBase64,
+        @RequestParam(value = "imgBase64Name", required = false) String imgBase64Name, HttpServletRequest request)
     {
         String filename = System.currentTimeMillis() + "_" + imgBase64Name;
-        if(file!=null)
+        if (file != null)
             filename = System.currentTimeMillis() + "_" + file.getOriginalFilename();
         String pathString = attachmentRoot + "/" + objectType + "/" + objectId + "/" + filename;
         File files = new File(pathString);
@@ -54,23 +55,32 @@ public class UploadController extends BaseController
         {
             files.getParentFile().mkdirs();
         }
-
+        
         try
         {
-            if (file != null){
+            if (file != null)
+            {
                 file.transferTo(files);
-            }else if(imgBase64!=null) {
+            }
+            else if (imgBase64 != null)
+            {
                 int index = imgBase64.indexOf(",");
                 imgBase64 = imgBase64.substring(index + 1);
                 generateImage(imgBase64, pathString);
             }
-
+            
             justImage(new File(pathString));
-
-            org.apache.commons.io.FileUtils.moveFile(new File(pathString + ".small.png"),
+            
+            if (new File(pathString + ".small.png").exists())
+            {
+                org.apache.commons.io.FileUtils.moveFile(new File(pathString + ".small.png"),
                     new File(pathString + ".small"));
-            org.apache.commons.io.FileUtils.moveFile(new File(pathString + ".middle.png"),
+            }
+            if (new File(pathString + ".middle.png").exists())
+            {
+                org.apache.commons.io.FileUtils.moveFile(new File(pathString + ".middle.png"),
                     new File(pathString + ".middle"));
+            }
         }
         catch (Exception e)
         {
@@ -136,11 +146,13 @@ public class UploadController extends BaseController
         
         return null;
     }
-
-    private static void generateImage(String realStr, String filePath) {
+    
+    private static void generateImage(String realStr, String filePath)
+    {
         BASE64Decoder decoder = new BASE64Decoder();
         BufferedOutputStream bos = null;
-        try {
+        try
+        {
             File f = new File(filePath);
             if (!f.getParentFile().exists())
                 f.getParentFile().mkdirs();
@@ -149,13 +161,20 @@ public class UploadController extends BaseController
             bos.write(data);
             bos.flush();
             bos.close();
-        } catch (IOException e) {
+        }
+        catch (IOException e)
+        {
             throw new RuntimeException(e.getMessage(), e);
-        } finally {
+        }
+        finally
+        {
             if (bos != null)
-                try {
+                try
+                {
                     bos.close();
-                } catch (IOException e) {
+                }
+                catch (IOException e)
+                {
                     e.printStackTrace();
                 }
         }
